Refactoring of infer and TDSL APIs. Forward towards great unification!
